microservices/statement-block-service/requirements.txt
flask==2.2.5
flask-cors==4.0.0
flask-limiter==3.5.0
anthropic==0.18.1
google-generativeai==0.8.3
python-dotenv==0.21.1
gunicorn==20.1.0
redis==4.5.5
pytest==7.2.0
pytest-flask==1.2.0